pyproject.toml
[project]
name = "new-personal-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"markdown>=3.7",
	"python-fasthtml",
	"python-frontmatter>=1.1.0",
]
